# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shradha Oil Center ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entities:
    get:
      operationId: listEntities
      tags: [entities]
      summary: List entities
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [customer, vendor, worker, salesman]
        - name: search
          in: query
          schema:
            type: string
        - name: mobile
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entity"

    post:
      operationId: createEntity
      tags: [entities]
      summary: Create entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/lookup:
    get:
      operationId: lookupEntityByMobile
      tags: [entities]
      summary: Lookup entity by mobile
      parameters:
        - name: mobile
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Entity or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityLookup"
  /entities/{id}:
    get:
      operationId: getEntity
      tags: [entities]
      summary: Get entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"

    patch:
      operationId: updateEntity
      tags: [entities]
      summary: Update entity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/{id}/ledger:
    get:
      operationId: getEntityLedger
      tags: [entities]
      summary: Get customer ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityLedger"
    post:
      operationId: createLedgerAdjustment
      tags: [entities]
      summary: Add a manual credit/debit adjustment to a customer ledger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LedgerAdjustmentInput"
      responses:
        "201":
          description: Created ledger entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerEntry"

  # INVOICES
...
components:
  schemas:
...
    Entity:
      type: object
      required: [id, type, name, mobile]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [customer, vendor, worker, salesman]
        name:
          type: string
        mobile:
          type: string
        gstin:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        pinCode:
          type: ["string", "null"]
        gpsLocation:
          type: ["string", "null"]
        pricingTier:
          type: ["string", "null"]
          enum: [retail, wholesale, null]
        outstandingBalance:
          type: number
        creditLimit:
          type: ["number", "null"]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
    EntityInput:
      type: object
      required: [type, mobile]
      description: |
        `name` is optional ONLY for retail customers (walk-ins) — if omitted, the server
        generates a label like "Retail Customer (98765 43210)". For wholesale customers,
        vendors, workers, and salesmen the name is required.
      properties:
        type:
          type: string
          enum: [customer, vendor, worker, salesman]
        name:
          type: string
        mobile:
          type: string
        gstin:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        district:
          type: string
        area:
          type: string
        pinCode:
          type: string
        gpsLocation:
          type: string
        pricingTier:
          type: string
          enum: [retail, wholesale]
        creditLimit:
          type: number
    EntityUpdate:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        mobile:
          type: string
        gstin:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        district:
          type: string
        area:
          type: string
        pinCode:
          type: string
        gpsLocation:
          type: string
        pricingTier:
          type: string
        creditLimit:
          type: number
    EntityLookup:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        entity:
          $ref: "#/components/schemas/Entity"
    EntityLedger:
      type: object
      required: [entity, outstandingBalance, entries]
      properties:
        entity:
          $ref: "#/components/schemas/Entity"
        outstandingBalance:
          type: number
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
    LedgerEntry:
      type: object
      required: [id, date, description, debit, credit, balance, type]
      properties:
        id:
          type: integer
        date:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
        balance:
          type: number
        type:
          type: string
          enum: [invoice, payment, adjustment]
        referenceId:
          type: ["integer", "null"]
        referenceNo:
          type: ["string", "null"]
        attachmentUrl:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
    LedgerAdjustmentInput:
      type: object
      required: [direction, amount]
      properties:
        direction:
          type: string
          enum: [debit, credit]
        amount:
          type: number
        notes:
          type: ["string", "null"]
        attachmentUrl:
          type: ["string", "null"]
...